Search book and select
