+++ client/src/component/_pathlessLayout/groups/$groupId/LinkIconButton.tsx
@@ -1,0 +1,4 @@
+import { IconButton } from "@mui/material";
+import { createLink } from "@tanstack/react-router";
+
+export const LinkIconButton = createLink(IconButton);
